requirements_direct.txt
# Direct dependencies for Atlassian MCP Server
fastapi==0.117.1
uvicorn==0.35.0
python-dotenv==1.1.1
requests==2.32.5
requests-oauthlib==2.0.0
Jinja2==3.1.6
python-multipart==0.0.20
starlette==0.48.0

# CrewAI and AI tools
crewai==0.130.0
crewai-tools==0.47.1
openai==1.99.9
langchain==0.3.27
langchain-openai==0.2.14

# MCP (Model Context Protocol)
mcp==1.14.0

# Data processing
pandas==2.3.2
numpy==2.3.3

# Additional utilities
PyYAML==6.0.2
